working pretty well now, still no love creating a libvirt VM though!
Foreman site comes up fine, DNS and DHCP integration appears to be
working. The Compute Resource for libvirt passes the web based test
but when I save a host, I just stay on the edit screen.

Going to switch gears and try to use oVirt instead, since
that is what my ISP will be using.